+++ s3datarepository/s3datarepository_test.go
@@ -34,10 +34,10 @@ func newMockS3Client(t *testing.T) s3iface.S3API {
 }
 
 func (m *mockS3Client) HeadBucketWithContext(ctx context.Context, input *s3.HeadBucketInput, opts ...request.Option) (*s3.HeadBucketOutput, error) {
-	if m.err["HeadBucketWithContext"] != nil {
+	if err, ok := m.err["HeadBucketWithContext"]; ok {
 		if m.headCount == 0 {
 			m.headCount++
-			return nil, m.err["HeadBucketWithContext"]
+			return nil, err
 		}
 	}
 
@@ -57,36 +57,36 @@ func (m *mockS3Client) HeadBucketWithContext(ctx context.Context, input *s3.Head
 }
 
 func (m *mockS3Client) CreateBucketWithContext(ctx context.Context, input *s3.CreateBucketInput, opts ...request.Option) (*s3.CreateBucketOutput, error) {
-	if m.err["CreateBucketWithContext"] != nil {
-		return nil, m.err["CreateBucketWithContext"]
+	if err, ok := m.err["CreateBucketWithContext"]; ok {
+		return nil, err
 	}
 	return &s3.CreateBucketOutput{Location: aws.String("/testbucket")}, nil
 }
 
 func (m *mockS3Client) DeleteBucketWithContext(ctx context.Context, input *s3.DeleteBucketInput, opts ...request.Option) (*s3.DeleteBucketOutput, error) {
-	if m.err["DeleteBucketWithContext"] != nil {
-		return nil, m.err["DeleteBucketWithContext"]
+	if err, ok := m.err["DeleteBucketWithContext"]; ok {
+		return nil, err
 	}
 	return &s3.DeleteBucketOutput{}, nil
 }
 
 func (m *mockS3Client) PutBucketEncryptionWithContext(ctx context.Context, input *s3.PutBucketEncryptionInput, opts ...request.Option) (*s3.PutBucketEncryptionOutput, error) {
-	if m.err["PutBucketEncryptionWithContext"] != nil {
-		return nil, m.err["PutBucketEncryptionWithContext"]
+	if err, ok := m.err["PutBucketEncryptionWithContext"]; ok {
+		return nil, err
 	}
 	return &s3.PutBucketEncryptionOutput{}, nil
 }
 
 func (m *mockS3Client) PutBucketTaggingWithContext(ctx context.Context, input *s3.PutBucketTaggingInput, opts ...request.Option) (*s3.PutBucketTaggingOutput, error) {
-	if m.err["PutBucketTaggingWithContext"] != nil {
-		return nil, m.err["PutBucketTaggingWithContext"]
+	if err, ok := m.err["PutBucketTaggingWithContext"]; ok {
+		return nil, err
 	}
 	return &s3.PutBucketTaggingOutput{}, nil
 }
 
 func (m *mockS3Client) PutPublicAccessBlockWithContext(ctx context.Context, input *s3.PutPublicAccessBlockInput, opts ...request.Option) (*s3.PutPublicAccessBlockOutput, error) {
-	if m.err["PutPublicAccessBlockWithContext"] != nil {
-		return nil, m.err["PutPublicAccessBlockWithContext"]
+	if err, ok := m.err["PutPublicAccessBlockWithContext"]; ok {
+		return nil, err
 	}
 	return &s3.PutPublicAccessBlockOutput{}, nil
 }
@@ -102,7 +102,6 @@ func TestNewDefaultRepository(t *testing.T) {
 	}
 
 	s, err := NewDefaultRepository(testConfig)
-
 	if err != nil {
 		t.Errorf("expected nil error, got: %s", err)
 	}
@@ -186,7 +185,7 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test success, with tags
-	s.S3.(*mockS3Client).headCount = 0
+	s = S3Repository{S3: newMockS3Client(t)}
 	s.S3.(*mockS3Client).err["HeadBucketWithContext"] = awserr.New("NotFound", "bucket not found", nil)
 
 	err := s.Provision(context.TODO(), "68004EEC-6044-45C9-91E5-AF836DCD9234", testTags)
@@ -195,7 +194,7 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test success, without tags
-	s.S3.(*mockS3Client).headCount = 0
+	s = S3Repository{S3: newMockS3Client(t)}
 	s.S3.(*mockS3Client).err["HeadBucketWithContext"] = awserr.New("NotFound", "bucket not found", nil)
 
 	err = s.Provision(context.TODO(), "68004EEC-6044-45C9-91E5-AF836DCD9234", []*dataset.Tag{})
@@ -204,10 +203,10 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test empty id
+	s = S3Repository{S3: newMockS3Client(t)}
 	id = ""
 	expectedCode = apierror.ErrBadRequest
 	expectedMessage = "invalid input"
-	s.S3.(*mockS3Client).err = make(map[string]error)
 
 	err = s.Provision(context.TODO(), id, testTags)
 	if aerr, ok := err.(apierror.Error); ok {
@@ -222,6 +221,7 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test existing id
+	s = S3Repository{S3: newMockS3Client(t)}
 	id = "68004EEC-6044-45C9-91E5-AF836DCD9234-exists"
 	expectedCode = apierror.ErrConflict
 	expectedMessage = "s3 bucket already exists"
@@ -239,11 +239,10 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test bucket create failure
+	s = S3Repository{S3: newMockS3Client(t)}
 	id = "68004EEC-6044-45C9-91E5-AF836DCD9234"
 	expectedCode = apierror.ErrServiceUnavailable
 	expectedMessage = fmt.Sprintf("failed to create s3 bucket dataset-%s", id)
-	s.S3.(*mockS3Client).headCount = 0
-	s.S3.(*mockS3Client).err = make(map[string]error)
 	s.S3.(*mockS3Client).err["HeadBucketWithContext"] = awserr.New("NotFound", "bucket not found", nil)
 	s.S3.(*mockS3Client).err["CreateBucketWithContext"] = awserr.New("InternalError", "Internal Error", nil)
 
@@ -260,10 +259,10 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test bucket create timeout failure
+	s = S3Repository{S3: newMockS3Client(t)}
 	id = "68004EEC-6044-45C9-91E5-AF836DCD9234-missing"
 	expectedCode = apierror.ErrInternalError
 	expectedMessage = fmt.Sprintf("failed to create bucket dataset-%s, timeout waiting for create: s3 bucket (dataset-%s) doesn't exist", id, id)
-	s.S3.(*mockS3Client).err = make(map[string]error)
 
 	err = s.Provision(context.TODO(), id, testTags)
 	if err == nil {
@@ -282,11 +281,10 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test bucket block public access failure
+	s = S3Repository{S3: newMockS3Client(t)}
 	id = "68004EEC-6044-45C9-91E5-AF836DCD9234"
 	expectedCode = apierror.ErrServiceUnavailable
 	expectedMessage = fmt.Sprintf("failed block public access for s3 bucket dataset-%s", id)
-	s.S3.(*mockS3Client).headCount = 0
-	s.S3.(*mockS3Client).err = make(map[string]error)
 	s.S3.(*mockS3Client).err["HeadBucketWithContext"] = awserr.New("NotFound", "bucket not found", nil)
 	s.S3.(*mockS3Client).err["PutPublicAccessBlockWithContext"] = awserr.New("InternalError", "Internal Error", nil)
 
@@ -307,11 +305,10 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test bucket enable encryption failure
+	s = S3Repository{S3: newMockS3Client(t)}
 	id = "68004EEC-6044-45C9-91E5-AF836DCD9234"
 	expectedCode = apierror.ErrServiceUnavailable
 	expectedMessage = fmt.Sprintf("failed to enable encryption for s3 bucket dataset-%s", id)
-	s.S3.(*mockS3Client).headCount = 0
-	s.S3.(*mockS3Client).err = make(map[string]error)
 	s.S3.(*mockS3Client).err["HeadBucketWithContext"] = awserr.New("NotFound", "bucket not found", nil)
 	s.S3.(*mockS3Client).err["PutBucketEncryptionWithContext"] = awserr.New("InternalError", "Internal Error", nil)
 
@@ -332,11 +329,10 @@ func TestProvision(t *testing.T) {
 	}
 
 	// test bucket tagging failure
+	s = S3Repository{S3: newMockS3Client(t)}
 	id = "68004EEC-6044-45C9-91E5-AF836DCD9234"
 	expectedCode = apierror.ErrServiceUnavailable
 	expectedMessage = fmt.Sprintf("failed to tag s3 bucket dataset-%s", id)
-	s.S3.(*mockS3Client).headCount = 0
-	s.S3.(*mockS3Client).err = make(map[string]error)
 	s.S3.(*mockS3Client).err["HeadBucketWithContext"] = awserr.New("NotFound", "bucket not found", nil)
 	s.S3.(*mockS3Client).err["PutBucketTaggingWithContext"] = awserr.New("InternalError", "Internal Error", nil)
 
@@ -374,6 +370,7 @@ func TestDelete(t *testing.T) {
 	}
 
 	// test empty id
+	s = S3Repository{S3: newMockS3Client(t)}
 	err = s.Delete(context.TODO(), "")
 	if aerr, ok := err.(apierror.Error); ok {
 		if aerr.Code != apierror.ErrBadRequest {
@@ -384,6 +381,7 @@ func TestDelete(t *testing.T) {
 	}
 
 	// test ErrCodeNoSuchBucket
+	s = S3Repository{S3: newMockS3Client(t)}
 	s.S3.(*mockS3Client).err["DeleteBucketWithContext"] = awserr.New(s3.ErrCodeNoSuchBucket, "bucket not found", nil)
 	err = s.Delete(context.TODO(), "68004EEC-6044-45C9-91E5-AF836DCD9234")
 	if aerr, ok := err.(apierror.Error); ok {
@@ -395,6 +393,7 @@ func TestDelete(t *testing.T) {
 	}
 
 	// test NotFound
+	s = S3Repository{S3: newMockS3Client(t)}
 	s.S3.(*mockS3Client).err["DeleteBucketWithContext"] = awserr.New("NotFound", "bucket not found", nil)
 	err = s.Delete(context.TODO(), "68004EEC-6044-45C9-91E5-AF836DCD9234")
 	if aerr, ok := err.(apierror.Error); ok {
@@ -406,6 +405,7 @@ func TestDelete(t *testing.T) {
 	}
 
 	// test BucketNotEmpty
+	s = S3Repository{S3: newMockS3Client(t)}
 	s.S3.(*mockS3Client).err["DeleteBucketWithContext"] = awserr.New("BucketNotEmpty", "bucket not empty", nil)
 	err = s.Delete(context.TODO(), "68004EEC-6044-45C9-91E5-AF836DCD9234")
 	if aerr, ok := err.(apierror.Error); ok {
@@ -417,6 +417,7 @@ func TestDelete(t *testing.T) {
 	}
 
 	// test non-aws error
+	s = S3Repository{S3: newMockS3Client(t)}
 	s.S3.(*mockS3Client).err["DeleteBucketWithContext"] = errors.New("things blowing up")
 	err = s.Delete(context.TODO(), "68004EEC-6044-45C9-91E5-AF836DCD9234")
 	if aerr, ok := err.(apierror.Error); ok {
